Usunięcie pakietu utils i przeniesienie istniejącego tam kodu do pakietu common
