+++ src/main/java/pl/umk/mat/gobooks/common/exceptions/ErrorResponse.java
@@ -1,0 +1,20 @@
+package pl.umk.mat.gobooks.common.exceptions;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+public class ErrorResponse {
+    private String message;
+    private List<String> details;
+
+    public ErrorResponse(String message, List<String> details) {
+        super();
+        this.message = message;
+        this.details = details;
+    }
+
+}
