+++ src/main/java/pl/umk/mat/gobooks/common/exceptions/ExceptionHandlerImpl.java
@@ -1,0 +1,70 @@
+package pl.umk.mat.gobooks.common.exceptions;
+
+import org.jetbrains.annotations.NotNull;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.ObjectError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.context.request.WebRequest;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@SuppressWarnings({"unchecked", "rawtypes"})
+@ControllerAdvice
+public class ExceptionHandlerImpl extends ResponseEntityExceptionHandler {
+    @ExceptionHandler(Exception.class)
+    public final ResponseEntity<Object> handleAllExceptions(Exception ex, WebRequest request) {
+        List<String> details = new ArrayList<>();
+        details.add(ex.getLocalizedMessage());
+        ErrorResponse error = new ErrorResponse("Server Error", details);
+        return new ResponseEntity(error, HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    @ExceptionHandler(BadRequest.class)
+    public final ResponseEntity<Object> handleBadRequestException(BadRequest ex, WebRequest request) {
+        List<String> details = new ArrayList<>();
+        details.add(ex.getLocalizedMessage());
+        ErrorResponse error = new ErrorResponse("Bad Request", details);
+        return new ResponseEntity(error, HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler(ResourceNotFound.class)
+    public final ResponseEntity<Object> handleResourceNotFoundException(ResourceNotFound ex, WebRequest request) {
+        List<String> details = new ArrayList<>();
+        details.add(ex.getLocalizedMessage());
+        ErrorResponse error = new ErrorResponse("Resource Not Found", details);
+        return new ResponseEntity(error, HttpStatus.NOT_FOUND);
+    }
+
+    @ExceptionHandler(ResourceAlreadyExist.class)
+    public final ResponseEntity<Object> handleResourceAlreadyExistException(ResourceAlreadyExist ex, WebRequest request) {
+        List<String> details = new ArrayList<>();
+        details.add(ex.getLocalizedMessage());
+        ErrorResponse error = new ErrorResponse("Resource Already Exist", details);
+        return new ResponseEntity(error, HttpStatus.CONFLICT);
+    }
+
+    @ExceptionHandler(Unauthorized.class)
+    public final ResponseEntity<Object> handleUnauthorizedException(Unauthorized ex, WebRequest request) {
+        List<String> details = new ArrayList<>();
+        details.add(ex.getLocalizedMessage());
+        ErrorResponse error = new ErrorResponse("Unauthorized", details);
+        return new ResponseEntity(error, HttpStatus.UNAUTHORIZED);
+    }
+
+    @NotNull
+    @Override
+    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex, HttpHeaders headers, HttpStatus status, WebRequest request) {
+        List<String> details = new ArrayList<>();
+        for (ObjectError error : ex.getBindingResult().getAllErrors()) {
+            details.add(error.getDefaultMessage());
+        }
+        ErrorResponse error = new ErrorResponse("Validation Failed", details);
+        return new ResponseEntity(error, HttpStatus.BAD_REQUEST);
+    }
+}
